+++ vNES/src/vnes/Mapper105.java
@@ -44,7 +44,7 @@ public class Mapper105 extends MapperDefault {
         super.mapperInternalStateSave(buf);
 
         // Version:
-        buf.putByte((int) 1);
+        buf.putByte( 1);
 
         // State:
         buf.putInt(irq_counter);
